Fixed arc drawing bug
The bug was caused by change of how current_point was handled:
previously Vector2D object of current_point was overwriten with
new Vector2D, but after changes it ended up being modified in place.
It caused references in region statements to point to current_point
changed by move_pointer instead of pointing to values passed at creation
time. That was fixed by restoring overwriting behaviour. Note that
for now values comeing from tokens has to be check for being None and
replaced with current_point or 0, depending on context.